requirements.txt
discord-py-interactions==5.5.1
python-dotenv==0.20.0
aiocache==0.12.1
colorthief==0.2.1
jurigged==0.5.6
psutil==5.9.6
humanize==4.8.0
python-dateutil==2.8.2
